Refactored content coverage reporting. Ignores test/action statements.
