+++ sample/coverage.json
@@ -1,8 +1,8 @@
 {
   "name": "Doc Detective Content Coverage Report",
-  "timestamp": "20221017-083214",
+  "timestamp": "20221019-152141",
   "summary": {
-    "covered": 12,
+    "covered": 11,
     "uncovered": 0,
     "onscreenText": {
       "covered": 2,
@@ -37,14 +37,13 @@
       "uncovered": 0
     },
     "interaction": {
-      "covered": 2,
+      "covered": 1,
       "uncovered": 0
     }
   },
   "files": [
     {
-      "file": "/config/workspace/doc-detective/sample/doc-content.md",
-      "covered": 12,
+      "covered": 11,
       "uncovered": 0,
       "onscreenText": {
         "covered": 2,
@@ -87,7 +86,7 @@
         "uncoveredMatches": []
       },
       "interaction": {
-        "covered": 2,
+        "covered": 1,
         "uncovered": 0,
         "uncoveredMatches": []
       }
